Extract standalone start in own module
